+++ src/SepomexPhp/LocationType.php
@@ -6,19 +6,16 @@ namespace SepomexPhp;
 
 use SepomexPhp\Traits\PropertyIdIntegerTrait;
 use SepomexPhp\Traits\PropertyNameStringTrait;
-use SepomexPhp\Traits\PropertyStateTrait;
 
-class District
+class LocationType
 {
     use PropertyIdIntegerTrait;
     use PropertyNameStringTrait;
-    use PropertyStateTrait;
 
-    public function __construct(int $id, string $name, State $state = null)
+    public function __construct(int $id, string $name)
     {
         $this->setId($id);
         $this->setName($name);
-        $this->setState($state);
     }
 
     public function asArray(): array
@@ -26,7 +23,6 @@ class District
         return [
             'id' => $this->id(),
             'name' => $this->name(),
-            'state' => $this->hasState() ? $this->state()->asArray() : null,
         ];
     }
 }